Move `processedDefault.js` file to `common` folder, leaving in `config` only files set by user
