+++ openmctStaticServer/index.html
@@ -10,7 +10,7 @@
     <script src="dictionary-plugin.js"></script>
     <script src="historical-telemetry-plugin.js"></script>
     <script src="realtime-telemetry-plugin.js"></script>
-    <script src="config/processedDefault.json"></script>
+    <script src="common/processedDefault.json"></script>
     <link rel="stylesheet" href="openmctdist/espressoTheme.css">
 </head>
 <body>
